+++ swift-sdk/Internal/InternalIterableAPI.swift
@@ -117,7 +117,7 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         _payloadData = data
     }
     
-    func setEmail(_ email: String?, authToken: String? = nil) {
+    func setEmail(_ email: String?, authToken: String? = nil, successHandler: OnSuccessHandler? = nil, failureHandler: OnFailureHandler? = nil) {
         ITBInfo()
         
         if _email == email && email != nil && authToken != nil {
@@ -133,13 +133,15 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         
         _email = email
         _userId = nil
+        _successCallback = successHandler
+        _failureCallback = failureHandler
         
         storeIdentifierData()
         
         onLogin(authToken)
     }
     
-    func setUserId(_ userId: String?, authToken: String? = nil) {
+    func setUserId(_ userId: String?, authToken: String? = nil, successHandler: OnSuccessHandler? = nil, failureHandler: OnFailureHandler? = nil) {
         ITBInfo()
         
         if _userId == userId && userId != nil && authToken != nil {
@@ -155,6 +157,8 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         
         _email = nil
         _userId = userId
+        _successCallback = successHandler
+        _failureCallback = failureHandler
         
         storeIdentifierData()
         
@@ -173,6 +177,7 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         guard let appName = pushIntegrationName else {
             let errorMessage = "Not registering device token - appName must not be nil"
             ITBError(errorMessage)
+            _failureCallback?(errorMessage, nil)
             onFailure?(errorMessage, nil)
             return
         }
@@ -187,8 +192,15 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
                                                   sdkVersion: localStorage.sdkVersion)
         requestHandler.register(registerTokenInfo: registerTokenInfo,
                                 notificationStateProvider: notificationStateProvider,
-                                onSuccess: onSuccess,
-                                onFailure: onFailure)
+                                onSuccess: { (_ data: [AnyHashable: Any]?) in
+                                                self._successCallback?(data)
+                                                onSuccess?(data)
+                                },
+                                onFailure: { (_ reason: String?, _ data: Data?) in
+                                                self._failureCallback?(reason, data)
+                                                onFailure?(reason, data)
+                                }
+        )
     }
     
     @discardableResult
@@ -434,6 +446,9 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
     private var _email: String?
     private var _payloadData: [AnyHashable: Any]?
     private var _userId: String?
+    private var _successCallback: OnSuccessHandler? = nil
+    private var _failureCallback: OnFailureHandler? = nil
+
     
     /// the hex representation of this device token
     private var hexToken: String?
@@ -543,6 +558,8 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         
         if config.autoPushRegistration {
             notificationStateProvider.registerForRemoteNotifications()
+        } else {
+            _successCallback?([:])
         }
         
         _ = inAppManager.scheduleSync()